fix(tests): Move all unused tests to excluded folder
